requirements.txt
# pip install -r requirements.txt
xlrd==1.2.0
pytorch>=1.2.0
torchvision>=0.4.0
numpy>=1.18.5
flask==2.2.3
flask-cors==3.0.10
uwsgi